+++ OptunaTests/Storage/RDB/StorageTests.cs
@@ -9,19 +9,6 @@ using Xunit;
 
 namespace Optuna.Storage.RDB.Tests
 {
-    public class CreateStorage
-    {
-        public CreateStorage()
-        {
-            string path = @"TestFile/created.db";
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
-            _ = new SqliteStorage(path, true);
-        }
-    }
-
     public class SqliteStorageTests : IClassFixture<CreateStorage>, IDisposable
     {
         private readonly string _createFilePath = @"TestFile/created.db";
